tests/requirements.txt
pytest==6.2.2
pytest-asyncio==0.14.0
pytest-aiohttp==0.3.0